Created a class User, with constructor which takes two parameters, username and password.
Created a method that creates and returns an array of User-objects with dedicated paramters.

The Main method now sets the return value of CreateUsers to an users array.
Then running the methods for the program.

Also added assignment text for accessability.
Created a method for the login process which checks user input towards the array of users and returns an 'active' User.

Created a method with the active /current user as parameter to run the menu program.