Adjusted RunMenu method. Now takes in two additional bool parameters that serve as a check for user to log in again or completely exit program. The method now returns true or false.
Moved sections of the code to method codeblocks, such as "presentable name" will ensure that the system will greet the user with a capital first letter regardless of input.

Greeting message moved to play at end of LogIn function

Adjusted the mock-bank alternatives to the ones dictated in the assignment.

Changed password from string to 4 digit pincode.

Added a method for welcoming the user when the program first runs.