Added a new class: BankFunctions - holds methods for specific account managements
Moved WelcomeMsg to MessageInformation class and called it.
Removed a redundant methodcall, User.CreateUsers.

Added a property balances to Users. Also added a third parameter to the constructor User. It take a double- array.
Populated the array in Create Users Method.

Moved a few more methods to MessageInformation class.
Adjusted the methodcalls so that they can be called from another class.