Completed project, v0.9 Updated Readme
For BankFunctions:
Moved order of the methods
RunMenu now contains all methods in the respective menu choice.

Populate accounts is created to iterate through a user[] and set values of type double to each account for each user.

Created a method ViewAccount is created that iterates and lists the active users accounts and their corresponding balance.

Created a method ChooseAccount that ensure that the user can only choose one of 1-3 accounts.

A method ViewBalance was created that console logs the balances for the accounts of the active user.

A TransferFunds account was created that takes the input from the user of which accounts to and from, and the amount that should be transfered - then executes the transfer if there are funds.

WithdrawFunds method was created that allows a user to choose and account and amount to withdraw. This needs to be confirmed with a pin, the sum is then subtracted from the account if there are sufficient funds.

In Users:
Adjusted the comments
Added another parameter accountNames.
Adjusted the constructor to accomade for the new parameter.

The CreateUsers method is adjusted accordingly and sets the Account names to each instanced user.

In method LogIn
Moved the user input and prompt to inside the while loop so that the user can try again if they wrote wrong.

RunMenu was moved from Users to BankFunctions.